pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "tefas-crawler>=0.5.0",
    "yfinance>=0.2.66",
]
